Bring all code in to getDataObject()
Moved all code in to the one function. this hurts my sole as some one who like to have many functions which do very specific things. Tough project to learn javascript on due to the order with which functions process. Language is very much designed for front end and to process simple backend calls.